Move ResponseDelivery components to Settings
