+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/settings/components/SetResponseDelivery.tsx
@@ -7,15 +7,15 @@ import { useSession } from "next-auth/react";
 
 import { useRefresh } from "@lib/hooks";
 import { isValidGovEmail } from "@lib/validation";
-import { ResponseEmail } from "./ResponseEmail";
-import { Radio } from "./shared";
+import { ResponseEmail } from "../../../components/ResponseEmail";
+import { Radio } from "../../../components/shared";
 import { Button } from "@clientComponents/globals";
 import { useTemplateApi } from "@lib/hooks/form-builder";
 import { useTemplateStore } from "@lib/store";
 import { completeEmailAddressRegex } from "@lib/utils/form-builder";
-import { toast } from "./shared/Toast";
+import { toast } from "../../../components/shared/Toast";
 import { ResponseDeliveryHelpButton } from "@formBuilder/components/shared";
-import { ClassificationType, ClassificationSelect } from "./ClassificationSelect";
+import { ClassificationType, ClassificationSelect } from "../../../components/ClassificationSelect";
 
 enum DeliveryOption {
   vault = "vault",
