+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/settings/components/SettingsNavigation.tsx
