+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/settings/layout.tsx
@@ -1,7 +1,7 @@
 import { serverTranslation } from "@i18n";
 import { LoggedOutTab, LoggedOutTabName } from "@serverComponents/form-builder/LoggedOutTab";
 import { auth } from "@lib/auth";
-import { SettingsNavigation } from "@formBuilder/[id]/settings/SettingsNavigation";
+import { SettingsNavigation } from "@/app/(gcforms)/[locale]/(form administration)/form-builder/[id]/settings/components/SettingsNavigation";
 
 export default async function Layout({
   children,
