+++ app/(gcforms)/[locale]/(form administration)/form-builder/components/index.ts
@@ -1,4 +1,3 @@
-export { ResponseDelivery } from "./ResponseDelivery";
 export { ShareModal } from "./ShareModal";
 export { ShareModalUnauthenticated } from "./ShareModalUnauthenticated";
 export { FormOwnership } from "./FormOwnership";
